refactor: presentation ui 모델은 presentation model 패키지 안에 넣기
