+++ android/app/src/main/java/com/woowacourse/ody/presentation/address/AddressSearchDialog.kt
@@ -15,7 +15,7 @@ import androidx.webkit.WebViewAssetLoader
 import com.woowacourse.ody.data.remote.location.repository.KakaoGeoLocationRepository
 import com.woowacourse.ody.databinding.DialogAddressSearchBinding
 import com.woowacourse.ody.presentation.address.listener.AddressReceiveListener
-import com.woowacourse.ody.presentation.address.ui.toGeoLocationUiModel
+import com.woowacourse.ody.presentation.address.model.toGeoLocationUiModel
 import com.woowacourse.ody.presentation.address.web.AddressSearchInterface
 import com.woowacourse.ody.presentation.address.web.LocalContentWebViewClient
