+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingroom/BindingAdapter.kt
@@ -4,7 +4,7 @@ import android.widget.TextView
 import androidx.databinding.BindingAdapter
 import com.woowacourse.ody.R
 import com.woowacourse.ody.domain.model.NotificationType
-import com.woowacourse.ody.presentation.meetingroom.uimodel.NotificationLogUiModel
+import com.woowacourse.ody.presentation.meetingroom.model.NotificationLogUiModel
 
 @BindingAdapter("setNotificationLogText")
 fun TextView.setNotificationLogText(log: NotificationLogUiModel) {
