+++ android/app/src/main/java/com/woowacourse/ody/presentation/meetingroom/adapter/NotificationLogViewHolder.kt
@@ -2,7 +2,7 @@ package com.woowacourse.ody.presentation.meetingroom.adapter
 
 import androidx.recyclerview.widget.RecyclerView
 import com.woowacourse.ody.databinding.ItemNotificationLogBinding
-import com.woowacourse.ody.presentation.meetingroom.uimodel.NotificationLogUiModel
+import com.woowacourse.ody.presentation.meetingroom.model.NotificationLogUiModel
 
 class NotificationLogViewHolder(private val binding: ItemNotificationLogBinding) :
     RecyclerView.ViewHolder(binding.root) {
