+++ android/app/src/main/res/layout/item_notification_log.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="log"
-            type="com.woowacourse.ody.presentation.meetingroom.uimodel.NotificationLogUiModel" />
+            type="com.woowacourse.ody.presentation.meetingroom.model.NotificationLogUiModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
